nr-ue-phy, nr-ue-rrc: Fix UE RSRP corruption from sample counter wrap
In multi-cell scenarios with data traffic, A3-triggered handovers stop permanently shortly after the applications start: every measurement report reaches the gNB with rsrpResult IE 0 for the serving cell and all neighbours, so the handover algorithm can never pick a target.

Three compounding defects in the UE measurement chain cause this:

1. NrUePhy::UeMeasurementsElement::rsrpNum is uint8_t, but PSS arrives with every DL CTRL slot (e.g. every 0.5 ms at numerology 1), i.e. hundreds of samples per cell per 200 ms filtering period. The counter wraps (400 samples -> 144), corrupting the average.
2. rsrpSum accumulates the samples in dBm although it is documented as linear, so the wrapped divisor turns the average into a huge negative dBm value that maps to reporting range 0, and deep-null sidelobe samples (PSS received while the gNB analog beam points at another UE) dominate the mean.
3. NrUeRrc::SaveUeMeasurements applies the layer-3 filter F = (1-a)F + aM in the dB domain with no finite guard, so a single -inf dBm sample (zero received power) poisons the stored value for the rest of the simulation.

Fix: widen the sample counters to uint32_t, accumulate RSRP in linear mW and convert to dBm only when reporting (skipping zero-power samples), and ignore non-finite RSRP values in the layer-3 filter. Update the expected RSRP values of the nr-ue-measurements-piecewise-1/-2 suites (transition-window reports shift under the corrected linear-domain averaging, steady-state values are unchanged) and extend the nr-handover-target simulation to 2 s, since the corrected serving-cell average now crosses the A2 threshold one report period later.

+++ model/nr-ue-phy.h
@@ -1023,11 +1023,15 @@ class NR_EXPORT NrUePhy : public NrPhy
     /// Summary results of measuring a specific cell. Used for layer-1 filtering.
     struct NR_EXPORT UeMeasurementsElement
     {
-        double rsrpSum;  //!< Sum of RSRP sample values in linear unit.
-        uint8_t rsrpNum; //!< Number of RSRP samples.
+        double rsrpSum; //!< Sum of RSRP sample values in linear unit.
+        // The counters must be wide enough for a full filtering period: PSS
+        // arrives with every DL CTRL slot (e.g. every 0.5 ms at mu=1), i.e.
+        // hundreds of samples per cell per period. A uint8_t counter wraps and
+        // corrupts the average (this silently disabled handovers).
+        uint32_t rsrpNum; //!< Number of RSRP samples.
         // For the moment rsrq is not supported so set to 0
-        double rsrqSum;  //!< Sum of RSRQ sample values in linear unit.
-        uint8_t rsrqNum; //!< Number of RSRQ samples.
+        double rsrqSum;   //!< Sum of RSRQ sample values in linear unit.
+        uint32_t rsrqNum; //!< Number of RSRQ samples.
     };
 
     /**
